# Remove redundant cache insert in resolve_cnames()